# apps/api/src/swagger.yaml
openapi: 3.0.3
info:
  title: AgriTrack Central API
  description: |
    AgriTrack Central Backend API for Smart CRM Machinery Monitoring.
    
    ## Features
    - **Machines**: Real-time machine monitoring and management
    - **Bookings**: Farmer booking system
    - **Analytics**: Dashboard statistics
    - **Notifications**: SMS, Push, and Email alerts
    - **Geofencing**: Location-based boundaries and breach detection
    - **Fuel Tracking**: Real-time fuel monitoring and theft detection
    - **Maintenance**: Scheduled maintenance management
    
    ## WebSocket Events
    Connect to `ws://localhost:3001` for real-time updates:
    - `device_update`: Real-time machine state
    - `booking_update`: Booking status changes
    - `initial_state`: All machines on connect
    
  version: 2.0.0
  contact:
    name: AgriTrack Team
    url: https://github.com/KD-3030/AgriTrack
servers:
  - url: http://localhost:3001
    description: Local development server
...
paths:
...
  /api/v1/notifications/preferences/{userId}:
    get:
      tags:
        - Notifications
      summary: Get user notification preferences
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User preferences
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPreferences'
  /api/v1/notifications/preferences:
    post:
      tags:
        - Notifications
      summary: Save notification preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationPreferences'
      responses:
        '200':
          description: Preferences saved
  /api/v1/notifications/push-token:
    post:
      tags:
        - Notifications
      summary: Register push notification token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                token:
                  type: string
                platform:
                  type: string
                  enum: [ios, android, web]
      responses:
        '200':
          description: Token registered
...
  /api/v1/maintenance/upcoming:
    get:
      tags:
        - Maintenance
      summary: Get upcoming maintenance
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 7
      responses:
        '200':
          description: Upcoming maintenance schedules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MaintenanceSchedule'
...
  /api/v1/maintenance/machine/{id}:
    get:
      tags:
        - Maintenance
      summary: Get machine maintenance schedules
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Machine maintenance schedules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MaintenanceSchedule'
  /api/v1/maintenance:
    post:
      tags:
        - Maintenance
      summary: Create maintenance schedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaintenanceRequest'
      responses:
        '201':
          description: Maintenance scheduled
  /api/v1/maintenance/{id}/complete:
    post:
      tags:
        - Maintenance
      summary: Mark maintenance as complete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                notes:
                  type: string
                cost:
                  type: number
      responses:
        '200':
          description: Maintenance completed
components:
  schemas:
...
    NotificationPreferences:
      type: object
      properties:
        userId:
          type: string
        smsEnabled:
          type: boolean
        pushEnabled:
          type: boolean
        emailEnabled:
          type: boolean
        alertTypes:
          type: array
          items:
            type: string
...
    MaintenanceSchedule:
      type: object
      properties:
        id:
          type: string
        machine_id:
          type: string
        type:
          type: string
        due_date:
          type: string
          format: date
        status:
          type: string
          enum: [scheduled, overdue, completed]
        notes:
          type: string
    MaintenanceRequest:
      type: object
      required:
        - machineId
        - type
        - dueDate
      properties:
        machineId:
          type: string
        type:
          type: string
        dueDate:
          type: string
          format: date
        recurring:
          type: boolean
        intervalDays:
          type: integer
        notes:
          type: string
...